Also move client index.js to be shared more easily
