+++ src/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "elm-cloudmodel-server",
+  "name": "elm-cloudmodel",
   "version": "0.0.1",
   "description": "Code for writing an Elm cloudmodel application",
-  "main": "server/index.js",
+  "main": "client/js/index.js",
   "scripts": {
     "dev": "nodemon src/server/index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
@@ -18,6 +18,7 @@
   },
   "dependencies": {
     "express": "^4.17.1",
-    "socket.io": "^2.2.0"
+    "socket.io": "^2.2.0",
+    "socket.io-client": "^2.2.0"
   }
 }
